requirements.txt
requests
python-amazon-sp-api
functions-framework

